REF: restructure things to avoid import side-effect.
